Rename module to task-manager.js
